+++ src/extension.ts
@@ -4,6 +4,7 @@ import I18nTranslateCompletionProvider from "./I18nTranslateCompletionProvider";
 import I18nTranslatePrefixCompletionProvider from "./I18nTranslatePrefixCompletionProvider";
 import I18nLocalizeCompletionProvider from "./I18nLocalizeCompletionProvider";
 import I18nHoverProvider from "./I18nHoverProvider";
+import I18nCodeActionProvider from "./I18nCodeActionProvider";
 
 const SELECTOR = ["ruby", "erb", "haml", "slim"];
 
@@ -46,6 +47,13 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.languages.registerHoverProvider(
       SELECTOR,
       new I18nHoverProvider(i18n)
+    )
+  );
+
+  context.subscriptions.push(
+    vscode.languages.registerCodeActionsProvider(
+      SELECTOR,
+      new I18nCodeActionProvider(i18n)
     )
   );
 }